+++ endFuncs.h
@@ -1,0 +1,6 @@
+#ifndef END_FUNCS_H_
+#define END_FUNCS_H_
+
+void shoutdown(void);
+
+#endif // END_FUNCS_H_
