+++ initFuncs.h
@@ -1,0 +1,13 @@
+#ifndef INIT_FUNCS_H_
+#define INIT_FUNCS_H_
+
+
+void printIntVal(int);
+
+bool initialize(void);
+
+Ball makeBall(int size);
+
+Player makePlayer(void);
+
+#endif // INIT_FUNCS_H_
